# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ClarifAI - AI Tutoring Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DocumentFileType = typeof DocumentFileType[keyof typeof DocumentFileType];


export const DocumentFileType = {
  pdf: 'pdf',
  docx: 'docx',
  txt: 'txt',
} as const;

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  processing: 'processing',
  ready: 'ready',
  error: 'error',
} as const;

export interface Document {
  id: number;
  title: string;
  /** @nullable */
  filename?: string | null;
  fileType: DocumentFileType;
  fileSize: number;
  status: DocumentStatus;
  chunkCount: number;
  /** @nullable */
  topic?: string | null;
  createdAt: string;
}

export type DocumentInputFileType = typeof DocumentInputFileType[keyof typeof DocumentInputFileType];


export const DocumentInputFileType = {
  pdf: 'pdf',
  docx: 'docx',
  txt: 'txt',
} as const;

export interface DocumentInput {
  title: string;
  filename?: string;
  fileType: DocumentInputFileType;
  fileSize: number;
  topic?: string;
  content: string;
}

export interface DocumentStats {
  total: number;
  ready: number;
  processing: number;
  totalChunks: number;
  totalSize: number;
}

export interface Conversation {
  id: number;
  title: string;
  messageCount: number;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  documentId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  title: string;
  documentId?: number;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Source {
  documentTitle: string;
  /** @nullable */
  pageNumber?: number | null;
  /** @nullable */
  excerpt?: string | null;
  relevance: number;
}

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  sources?: Source[];
  createdAt: string;
}

export interface ConversationDetail {
  id: number;
  title: string;
  messages: Message[];
  /** @nullable */
  documentId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface MessageInput {
  content: string;
}

export type QuizDifficulty = typeof QuizDifficulty[keyof typeof QuizDifficulty];


export const QuizDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface Quiz {
  id: number;
  title: string;
  questionCount: number;
  difficulty: QuizDifficulty;
  /** @nullable */
  topic: string | null;
  /** @nullable */
  documentId?: number | null;
  /** @nullable */
  bestScore: number | null;
  attemptCount: number;
  createdAt: string;
}

export type QuizGenerationInputDifficulty = typeof QuizGenerationInputDifficulty[keyof typeof QuizGenerationInputDifficulty];


export const QuizGenerationInputDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface QuizGenerationInput {
  title: string;
  topic?: string;
  documentId?: number;
  questionCount: number;
  difficulty: QuizGenerationInputDifficulty;
}

export interface QuizQuestion {
  id: number;
  quizId: number;
  question: string;
  options: string[];
  correctAnswer: number;
  explanation: string;
}

export type QuizDetailDifficulty = typeof QuizDetailDifficulty[keyof typeof QuizDetailDifficulty];


export const QuizDetailDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface QuizDetail {
  id: number;
  title: string;
  questions: QuizQuestion[];
  difficulty: QuizDetailDifficulty;
  /** @nullable */
  topic: string | null;
  /** @nullable */
  documentId?: number | null;
  createdAt: string;
}

export interface QuizAttemptInput {
  answers: number[];
  timeSpent: number;
}

export interface QuestionResult {
  questionId: number;
  selectedAnswer: number;
  correctAnswer: number;
  isCorrect: boolean;
  explanation?: string;
}

export interface QuizResult {
  quizId: number;
  score: number;
  correctCount: number;
  totalQuestions: number;
  timeSpent: number;
  passed: boolean;
  questionResults?: QuestionResult[];
}

export interface FlashcardSet {
  id: number;
  title: string;
  cardCount: number;
  studiedCount: number;
  masteredCount: number;
  /** @nullable */
  topic: string | null;
  /** @nullable */
  documentId?: number | null;
  createdAt: string;
}

export interface FlashcardGenerationInput {
  title: string;
  topic?: string;
  documentId?: number;
  cardCount: number;
}

export interface Flashcard {
  id: number;
  setId: number;
  front: string;
  back: string;
  mastered: boolean;
  /** @nullable */
  difficulty?: string | null;
}

export interface FlashcardSetDetail {
  id: number;
  title: string;
  cards: Flashcard[];
  /** @nullable */
  topic: string | null;
  /** @nullable */
  documentId?: number | null;
  createdAt: string;
}

export interface FlashcardProgressUpdate {
  masteredCardIds: number[];
}

export interface DashboardStats {
  totalDocuments: number;
  totalConversations: number;
  totalQuizzes: number;
  totalFlashcardSets: number;
  averageQuizScore: number;
  studyStreak: number;
  totalStudyTime: number;
  weeklyMessages: number;
}

export interface ActivityEntry {
  date: string;
  messageCount: number;
  quizCount: number;
  flashcardCount: number;
}

export interface TopicStat {
  topic: string;
  count: number;
  percentage: number;
}

export interface QuizPerformanceEntry {
  date: string;
  averageScore: number;
  attemptCount: number;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  student: 'student',
  admin: 'admin',
} as const;

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  role: UserProfileRole;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  institution?: string | null;
  joinedAt: string;
}

export interface ProfileUpdate {
  name?: string;
  bio?: string;
  institution?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  student: 'student',
  admin: 'admin',
} as const;

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role: RegisterInputRole;
}

export interface AuthResponse {
  user: UserProfile;
  token: string;
}