# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ClarifAI - AI Tutoring Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all uploaded documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document for RAG processing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /quizzes:
    get:
      operationId: listQuizzes
      tags: [quizzes]
      summary: List all quizzes
      responses:
        "200":
          description: Quizzes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: generateQuiz
      tags: [quizzes]
      summary: Generate a quiz from uploaded documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizGenerationInput"
      responses:
        "201":
          description: Generated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get quiz with questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizDetail"
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete a quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quizzes/{id}/submit:
    post:
      operationId: submitQuizAttempt
      tags: [quizzes]
      summary: Submit a quiz attempt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAttemptInput"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"

  # ── Flashcards ───────────────────────────────────────────────────────────────
...
  /users/profile:
    get:
      operationId: getProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
components:
  schemas:
...
    Document:
      type: object
      required: [id, title, fileType, fileSize, status, chunkCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        filename:
          type: ["string", "null"]
        fileType:
          type: string
          enum: [pdf, docx, txt]
        fileSize:
          type: integer
        status:
          type: string
          enum: [processing, ready, error]
        chunkCount:
          type: integer
        topic:
          type: ["string", "null"]
        createdAt:
          type: string
    DocumentInput:
      type: object
      required: [title, fileType, fileSize, content]
      properties:
        title:
          type: string
        filename:
          type: string
        fileType:
          type: string
          enum: [pdf, docx, txt]
        fileSize:
          type: integer
        topic:
          type: string
        content:
          type: string
...
    Quiz:
      type: object
      required: [id, title, questionCount, difficulty, topic, bestScore, attemptCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        questionCount:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
        topic:
          type: ["string", "null"]
        documentId:
          type: ["integer", "null"]
        bestScore:
          type: ["number", "null"]
        attemptCount:
          type: integer
        createdAt:
          type: string
    QuizGenerationInput:
      type: object
      required: [title, questionCount, difficulty]
      properties:
        title:
          type: string
        topic:
          type: string
        documentId:
          type: integer
        questionCount:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
    QuizQuestion:
      type: object
      required: [id, quizId, question, options, correctAnswer, explanation]
      properties:
        id:
          type: integer
        quizId:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: integer
        explanation:
          type: string
    QuizDetail:
      type: object
      required: [id, title, questions, difficulty, topic, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        difficulty:
          type: string
          enum: [easy, medium, hard]
        topic:
          type: ["string", "null"]
        documentId:
          type: ["integer", "null"]
        createdAt:
          type: string
    QuizAttemptInput:
      type: object
      required: [answers, timeSpent]
      properties:
        answers:
          type: array
          items:
            type: integer
        timeSpent:
          type: integer
    QuizResult:
      type: object
      required: [quizId, score, correctCount, totalQuestions, timeSpent, passed]
      properties:
        quizId:
          type: integer
        score:
          type: number
        correctCount:
          type: integer
        totalQuestions:
          type: integer
        timeSpent:
          type: integer
        passed:
          type: boolean
        questionResults:
          type: array
          items:
            $ref: "#/components/schemas/QuestionResult"
    QuestionResult:
      type: object
      required: [questionId, selectedAnswer, correctAnswer, isCorrect]
      properties:
        questionId:
          type: integer
        selectedAnswer:
          type: integer
        correctAnswer:
          type: integer
        isCorrect:
          type: boolean
        explanation:
          type: string

    # ── Flashcards ─────────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [id, name, email, role, joinedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin]
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        institution:
          type: ["string", "null"]
        joinedAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        institution:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, admin]
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string