# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ClarifAI - AI Tutoring Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentFileType } from './documentFileType';
import type { DocumentStatus } from './documentStatus';

export interface Document {
  id: number;
  title: string;
  /** @nullable */
  filename?: string | null;
  fileType: DocumentFileType;
  fileSize: number;
  status: DocumentStatus;
  chunkCount: number;
  /** @nullable */
  topic?: string | null;
  createdAt: string;
}