# lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ClarifAI - AI Tutoring Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { QuizDifficulty } from './quizDifficulty';

export interface Quiz {
  id: number;
  title: string;
  questionCount: number;
  difficulty: QuizDifficulty;
  /** @nullable */
  topic: string | null;
  /** @nullable */
  documentId?: number | null;
  /** @nullable */
  bestScore: number | null;
  attemptCount: number;
  createdAt: string;
}